+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.2-Alpha
 ### Features
 - Add Neurosama family attack: if you attack any family member of Neurosama, you will be attacked by all family members.
+- Heart item can be eaten and level up player's absorption effect, up to absorption 5.
 
 ## 0.0.1-Alpha
 ### Features